worker/requirements.txt
nova-act>=0.1.0
boto3>=1.35.0
redis>=5.0.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
